+++ frontend/src/queries/useChartBlockMetrics.ts
@@ -31,6 +31,7 @@ const BlockMetricsQuery = gql<BlockMetricsQueryResponse>`
 
 export const useBlockMetricsQuery = (period: Ref<MetricsPeriod>) => {
 	const { data, executeQuery, fetching } = useQuery({
+		context: { url: useRuntimeConfig().public.apiUrlRust },
 		query: BlockMetricsQuery,
 		requestPolicy: 'cache-and-network',
 		variables: { period },