+++ backend-rust/src/graphql_api.rs
@@ -150,6 +150,8 @@ pub struct ApiServiceConfig {
         default_value = "100"
     )]
     token_holder_addresses_collection_limit: u64,
+    #[arg(long, env = "CCDSCAN_API_CONFIG_TOKEN_EVENTS_COLLECTION_LIMIT", default_value = "100")]
+    token_events_collection_limit: u64,
 }
 
 #[derive(MergedObject, Default)]