+++ frontend/CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ### Fix
 
+- Remove unused fields in query for cis2 token events.
 - Remove unused sub-query for account transactions in the accounts page.
 
 ## [1.6.0] - 2024-11-05